Modularizing SimpleProject
