--- src/Enums.ts
@@ -1,4 +1,0 @@
-enum ProjectStatus{
-    ACTIVE = 'active',
-    FINISHED = 'finished'
-}
