+++ src/decorators/autobind.ts
@@ -1,0 +1,16 @@
+//Decorator Autobind
+export function autoBind(
+    target: any, 
+    methodName: string, 
+    descriptor: PropertyDescriptor
+){
+    const originalMethod = descriptor.value;
+    const adjDescriptor: PropertyDescriptor = {
+        configurable: true,
+        get(){
+            const boundFn = originalMethod.bind(this);
+            return boundFn;
+        }
+    }
+    return adjDescriptor;
+}
